WIP: clean up
